+++ buildUtils/network/.null-ls_821720_runWhenBackOnline.zsh
@@ -1,2 +1,0 @@
-cd $ULLD_DEV_ROOT
-p add zod find-up ink-progress-bar ink-ascii ink-quicksearch-input ink-task-list mkdirp --filter @ulld/cli