+++ buildUtils/network/runWhenBackOnline.zsh
@@ -1,2 +1,1 @@
 cd $ULLD_DEV_ROOT
-p add zod find-up contains-path node-emoji ink-progress-bar ink-ascii ink-quicksearch-input ink-task-list mkdirp --filter @ulld/cli