+++ packages/full-form/src/inputs/reorder/general/reorderItemWithTrigger.tsx
@@ -6,17 +6,17 @@ import React, { ReactNode } from "react";
 export interface ReorderItemWithTriggerProps<ItemType extends unknown> {
     item: ItemType;
     classes?: {
-        dragHandle?: string
-        item?: string
-    }
-    label?: ReactNode
+        dragHandle?: string;
+        item?: string;
+    };
+    label?: ReactNode;
 }
 
 /* BUG: Fix this issue with the drag handle when on WIFI. Currently getting some context error from framer-motion. */
 const ReorderItemWithTrigger = <ItemType extends unknown>({
     item,
-classes = {},
-    label
+    classes = {},
+    label,
 }: ReorderItemWithTriggerProps<ItemType>) => {
     const controls = useDragControls();
     return (
@@ -27,10 +27,7 @@ classes = {},
             className={cn("border grid grid-cols-[1fr_32px] px-4 py-3", classes.item)}
         >
             {label}
-            <GripIcon
-                className={classes.dragHandle}
-                onPointerDown={controls.start}
-            />
+            <GripIcon className={classes.dragHandle} onPointerDown={controls.start} />
         </Reorder.Item>
     );
 };