+++ packages/navigation/src/pages/settingPage/footerSections/footerSectionItemLink.tsx
@@ -1,7 +1,9 @@
-import React from "react";
+import React, { PointerEvent } from "react";
 import { NavigationFormSettingSchema } from "../form/schema";
-import { Reorder, useDragControls } from "framer-motion";
-import { GripIcon, Trash2Icon } from "lucide-react";
+import { motion, Reorder, useDragControls } from "framer-motion";
+import { GripIcon as _GripIcon, Trash2Icon } from "lucide-react";
+
+const GripIcon = motion(_GripIcon)
 
 interface FooterSectionItemLinkProps {
     item: NavigationFormSettingSchema["footerSections"][number]["items"][number];
@@ -14,16 +16,20 @@ const FooterSectionItemLink = ({
     sectionIndex,
     itemIndex,
 }: FooterSectionItemLinkProps) => {
-    /* const controls = useDragControls(); */
+    const controls = useDragControls();
+    function startDrag(event: PointerEvent<Element> | PointerEvent) {
+        controls.start(event, { snapToCursor: true });
+    }
     return (
         <Reorder.Item
-            key={`${item.label}-${sectionIndex}`}
-            value={item}
             className={
-                "w-full px-3 py-2 border bg-background cursor-pointer hover:bg-secondary/60 transition-colors duration-200"
+                "w-full relative px-3 py-2 border bg-background hover:bg-secondary/60 transition-colors duration-200"
             }
-        /* dragListener={false} */
-        /* dragControls={controls} */
+            drag="y"
+            key={`${item.label}-linkItem`}
+            value={item}
+            dragListener={false}
+            dragControls={controls}
         >
             <div className={"grid grid-cols-[30px_auto_30px] place-items-center"}>
                 <div
@@ -37,8 +43,11 @@ const FooterSectionItemLink = ({
                 </div>
                 <div className={"w-full"}>{item.label}</div>
                 <GripIcon
-                    /* onPointerDown={(e) => controls.start(e)} */
-                    className={"w-3 h-3 text-muted-foreground"}
+                    onPointerDown={startDrag}
+                    className={"w-3 h-3 text-muted-foreground cursor-grab"}
+                    whileDrag={{
+                        cursor: "grabbing"
+                    }}
                 />
             </div>
         </Reorder.Item>