+++ packages/navigation/src/pages/settingPage/form/schema.tsx
@@ -1,4 +1,6 @@
 import { z } from "zod";
+import { faker } from "@faker-js/faker";
+
 
 export const navigationLinkSchema = z.object({
     label: z.string().max(12),
@@ -24,3 +26,66 @@ export const navigationFormSchemaWithUtilities = navigationFormSettingSchema.mer
 export type NavigationFormSettingSchema = z.infer<typeof navigationFormSettingSchema>
 
 export type NavigationFormWithUtilityFields = z.infer<typeof navigationFormSchemaWithUtilities>
+
+
+
+export const developmentDefaultValues: NavigationFormWithUtilityFields = {
+
+    footerSectionInput: "",
+    footerSections: [
+        {
+            label: faker.lorem.words({min: 1, max: 3}),
+            items: [
+                {
+                    label: faker.lorem.words({min: 1, max: 2}),
+                    href: "https://google.com",
+                },
+
+                {
+                    label: faker.lorem.words({min: 1, max: 2}),
+                    href: "https://google.com",
+                },
+                {
+                    label: faker.lorem.words({min: 1, max: 2}),
+                    href: "https://google.com",
+                },
+            ],
+        },
+        {
+            label: faker.lorem.words({min: 1, max: 3}),
+            items: [
+                {
+                    label: faker.lorem.words({min: 1, max: 2}),
+                    href: "https://google.com",
+                },
+
+                {
+                    label: faker.lorem.words({min: 1, max: 2}),
+                    href: "https://google.com",
+                },
+                {
+                    label: faker.lorem.words({min: 1, max: 2}),
+                    href: "https://google.com",
+                },
+            ],
+        },
+        {
+            label: faker.lorem.words({min: 1, max: 3}),
+            items: [
+                {
+                    label: faker.lorem.words({min: 1, max: 2}),
+                    href: "https://google.com",
+                },
+
+                {
+                    label: faker.lorem.words({min: 1, max: 2}),
+                    href: "https://google.com",
+                },
+                {
+                    label: faker.lorem.words({min: 1, max: 2}),
+                    href: "https://google.com",
+                },
+            ],
+        },
+    ],
+}