+++ packages/schemas/src/mdx/parsing/params/index.ts
@@ -1,1 +1,1 @@
-export * from "./main.ts"
+export * from "./main"