+++ packages/calendar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ulld/calendar
 
+## 0.0.124
+
+### Patch Changes
+
+- @ulld/developer@0.0.124
+
+## 0.0.123
+
+### Patch Changes
+
+- @ulld/developer@0.0.123
+
 ## 0.0.122
 
 ### Patch Changes