+++ packages/developer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ulld/developer
 
+## 0.0.124
+
+### Patch Changes
+
+- @ulld/configschema@0.0.124
+- @ulld/utilities@0.0.124
+
+## 0.0.123
+
+### Patch Changes
+
+- @ulld/configschema@0.0.123
+- @ulld/utilities@0.0.123
+
 ## 0.0.122
 
 ### Patch Changes