+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ulld/icons
 
+## 0.0.124
+
+### Patch Changes
+
+- @ulld/hooks@0.0.124
+- @ulld/utilities@0.0.124
+
+## 0.0.123
+
+### Patch Changes
+
+- @ulld/hooks@0.0.123
+- @ulld/utilities@0.0.123
+
 ## 0.0.122
 
 ### Patch Changes