+++ packages/math/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ulld/math
 
+## 0.0.124
+
+### Patch Changes
+
+- @ulld/hooks@0.0.124
+- @ulld/state@0.0.124
+- @ulld/icons@0.0.124
+- @ulld/configschema@0.0.124
+- @ulld/utilities@0.0.124
+
+## 0.0.123
+
+### Patch Changes
+
+- @ulld/hooks@0.0.123
+- @ulld/state@0.0.123
+- @ulld/icons@0.0.123
+- @ulld/configschema@0.0.123
+- @ulld/utilities@0.0.123
+
 ## 0.0.122
 
 ### Patch Changes