+++ packages/note-network/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ulld/note-network
 
+## 0.0.124
+
+### Patch Changes
+
+- @ulld/api@0.0.124
+- @ulld/configschema@0.0.124
+- @ulld/utilities@0.0.124
+- @ulld/developer@0.0.124
+
+## 0.0.123
+
+### Patch Changes
+
+- @ulld/api@0.0.123
+- @ulld/configschema@0.0.123
+- @ulld/utilities@0.0.123
+- @ulld/developer@0.0.123
+
 ## 0.0.122
 
 ### Patch Changes