+++ packages/whiteboard/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ulld/whiteboard
 
+## 0.0.124
+
+### Patch Changes
+
+- @ulld/api@0.0.124
+- @ulld/hooks@0.0.124
+- @ulld/state@0.0.124
+- @ulld/utilities@0.0.124
+- @ulld/developer@0.0.124
+
+## 0.0.123
+
+### Patch Changes
+
+- @ulld/api@0.0.123
+- @ulld/hooks@0.0.123
+- @ulld/state@0.0.123
+- @ulld/utilities@0.0.123
+- @ulld/developer@0.0.123
+
 ## 0.0.122
 
 ### Patch Changes